Feature/finish app

+++ app/src/main/java/com/example/weatherapp/data/service/NotificationHelper.kt
@@ -191,7 +191,7 @@ object NotificationHelper {
             .build()
     }
 
-    fun createWeatherNotification(context: Context, temp: Int, desc: String, city: String) {
+    fun createWeatherNotification(context: Context, temp: Double, desc: String, city: String) {
         val channelId = "weather_channel"
         val manager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
